requirements.txt
alabaster==0.7.13
asgiref==3.6.0
Babel==2.12.1
black==23.3.0
boto3==1.26.112
botocore==1.29.112
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect==1.2.0
defusedxml==0.7.1
distro==1.8.0
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-dynamic-fixture==3.1.2
django-environ==0.10.0
django-js-asset==2.0.0
django-rest-knox==4.2.0
django-simple-history==3.3.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.19
drf-yasg==1.21.5
Faker==18.4.0
gunicorn==20.1.0
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.0
pathspec==0.11.1
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.0
PyJWT==2.6.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
pytest==7.3.0
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
roman==4.0
ruamel.yaml==0.17.21
s3transfer==0.6.0
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Wand==0.6.11
